removed azure service runtime
adapted library to work outside of an azure web role instance. I remove
the windows azure service runtime assembly and the associated code in
the dependency resolver extension class. I replaced the code with a
simple retrieval of the machines local ip address as the instance number
in the GetRoleInstance method. Also updated all packages to their latest
releases. Due to the service bus sdk update from 1.6 to 1.7, i adapted
the code in the servicebusmessage bus to use the new
'CreateFromConnectionString' methods as preferred by the documentation.
Becuase of that I also changed the signature of the dependencey resolver
extension methods.